configure: use gcc over cc
We don't know what the system cc supports in terms of command line
flags, it could be tcc for all we know. Aiming for gcc over cc allows
the build to succeed when cc is not gcc or clang, but the user can
always override it themselves with the -c configure switch.