fs: init function "status"
fs.status returns a table containing information concerning the status
of a file. Currently, the table includes the following information:

  string  path       -- the given path to the file; the first argument
  integer mode       -- bitmask of file modes; see chmod(2)
  integer uid        -- owner of the file in UID form
  integer gid        -- owner group of the file in GID form
  integer accessdate -- date the file was last accessed
  integer modifydate -- date the file was last modified
  integer chdate     -- date the file's status last changed

Dates are given in the form of an integer number representing the
number of seconds since Jan 1, 1970 (unixtime).

For more information consult the definition of sys/stat.h
according to POSIX:
  https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/sys_stat.h.html